Add 1C docker

+++ dbmssql/dbproject/landing/audit/Functions/Template_LogProc.sql
@@ -2,7 +2,7 @@ CREATE FUNCTION [audit].[Template_LogProc](
 ) RETURNS @LogProc TABLE (
   [ID]            bigint        identity(1,1) NOT NULL,
   [LogID]         bigint        NOT NULL Primary Key,
-  [AuditTypeID]   tinyint       NOT NULL,
+  [AuditTypeID]   int           NOT NULL,
   [Msg]           varchar(max)  COLLATE Cyrillic_General_CI_AS NULL
 )
 AS
